Add low-level parsing API for binary IRAP
The API is a set of zero-allocation parsing routines that operates
purely on input and output buffers. In the form of raw void pointers.
The functions for reading the IRAP map is designed to support streaming
operations.

The API is public, but is primarily intended as a building block for a
higher level API to come.

The API is added in full, but most implementations are still missing.

I'm still evaluating if void* is the best choice for the buffers,
especially the map buffers which should always be floats. I will
re-evaluate once I've gained some more experience with the API.